JAMES-1646 move mailbox related model classes into a mailbox package
git-svn-id: https://svn.apache.org/repos/asf/james/project/trunk@1720372 13f79535-47bb-0310-9956-ffa450edef68
