+++ server/protocols/jmap/src/main/java/org/apache/james/jmap/model/GetMailboxesResponse.java
@@ -25,6 +25,7 @@ import org.apache.commons.lang.NotImplementedException;
 import org.apache.james.jmap.methods.Method;
 
 import com.google.common.collect.ImmutableList;
+import org.apache.james.jmap.model.mailbox.Mailbox;
 
 public class GetMailboxesResponse implements Method.Response {
