+++ server/protocols/jmap/src/main/java/org/apache/james/jmap/model/mailbox/Role.java
@@ -1,4 +1,3 @@
-package org.apache.james.jmap.model;
 /****************************************************************
  * Licensed to the Apache Software Foundation (ASF) under one   *
  * or more contributor license agreements.  See the NOTICE file *
@@ -17,40 +16,38 @@ package org.apache.james.jmap.model;
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.Test;
+package org.apache.james.jmap.model.mailbox;
 
 import java.util.Locale;
 import java.util.Optional;
 
-public class RoleTest {
+public enum Role {
 
-    @Test
-    public void fromShouldReturnInbox() {
-        assertThat(Role.from("inbox")).isEqualTo(Optional.of(Role.INBOX));
+    INBOX("inbox"),
+    ARCHIVE("archive"),
+    DRAFTS("drafts"),
+    OUTBOX("outbox"),
+    SENT("sent"),
+    TRASH("trash"),
+    SPAM("spam"),
+    TEMPLATES("templates");
+
+    private String name;
+
+    Role(String name) {
+        this.name = name;
     }
 
-    @Test
-    public void fromShouldReturnEmptyWhenUnknownValue() {
-        assertThat(Role.from("jjjj")).isEqualTo(Optional.empty());
-    }
-
-    @Test
-    public void fromShouldReturnInboxWhenContainsUppercaseValue() {
-        assertThat(Role.from("InBox")).isEqualTo(Optional.of(Role.INBOX));
-    }
-
-    @Test
-    public void fromShouldReturnInboxWhenContainsUppercaseValueInTurkish() {
-        Locale previousLocale = Locale.getDefault();
-        Locale.setDefault(Locale.forLanguageTag("tr"));
-        try {
-            assertThat(Role.from("InBox")).isEqualTo(Optional.of(Role.INBOX));
-        } finally {
-            Locale.setDefault(previousLocale);
+    public static Optional<Role> from(String name) {
+        for (Role role : values()) {
+            if (role.serialize().equals(name.toLowerCase(Locale.ENGLISH))) {
+                return Optional.of(role);
+            }
         }
+        return Optional.empty();
     }
 
+    public String serialize() {
+        return name;
+    }
 }
